Move data and fetching effect back to controller
